refactor: Factoring out more ops from executor (LdaConstant, Return)
