+++ src/machine/ops/index.js
@@ -1,7 +1,9 @@
 module.exports = {
   LdaSmi: require('./LdaSmi'),
+  LdaConstant: require('./LdaConstant'),
   LdaUndefined: require('./LdaUndefined'),
   LdaZero: require('./LdaZero'),
+  Return: require('./Return'),
   MulSmi: require('./MulSmi'),
   Star: require('./Star'),
   TestLessThan: require('./TestLessThan'),
